requirements.txt
aiosmtplib==5.1.0
annotated-types==0.7.0
anthropic==0.84.0
anyio==4.12.1
attrs==25.4.0
certifi==2026.2.25
cffi==2.0.0
click==8.3.1
cryptography==46.0.5
distro==1.9.0
docstring_parser==0.17.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
jiter==0.13.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
mcp==1.26.0
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
PyJWT==2.11.0
python-dotenv==1.2.2
python-multipart==0.0.22
referencing==0.37.0
rpds-py==0.30.0
sniffio==1.3.1
sse-starlette==3.3.2
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.41.0
